+++ docs/src/examples/components/Dropdown/Types/DropdownExampleClearable.shorthand.tsx
@@ -1,4 +1,4 @@
-import { Dropdown } from '@stardust-ui/react'
+import { Dropdown, Flex } from '@stardust-ui/react'
 import * as React from 'react'
 
 const inputItems = [
@@ -14,7 +14,16 @@ const inputItems = [
 ]
 
 const DropdownClearableExample = () => (
-  <Dropdown clearable items={inputItems} placeholder="Select your hero" />
+  <Flex column gap="gap.medium">
+    <Dropdown clearable items={inputItems} placeholder="Select your hero" />
+    <Dropdown
+      clearable
+      items={inputItems}
+      noResultsMessage="We couldn't find any matches."
+      placeholder="Select your hero"
+      search
+    />
+  </Flex>
 )
 
 export default DropdownClearableExample